feat: move editable Agent Instructions to Settings page
Relocate instruction cards and streamline Instructions page.

#VERCEL_SKIP
